# swagger_v1.yaml
openapi: 3.0.3
info:
  title: TravelGuide API
  description: API для управления городами и пользователями
  version: 2.0.0
servers:
  - url: http://localhost:8800/api/v2
    description: Development server
paths:
...
  /api/v1/entertainments:
    get:
      summary: Get all entertainments
      description: Retrieve all entertainments and render HTML page
      operationId: get_all_entertainments
      tags:
        - entertainments
      responses:
        '201':
          description: HTML page with entertainments
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal server error

    post:
      summary: Create new entertainment
      description: Create a new entertainment and return HTML response
      operationId: create_entertainment
      tags:
        - entertainments
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                city_id:
                  type: integer
                event_time:
                  type: string
                  format: date-time
                address:
                  type: string
                cost:
                  type: number
                description:
                  type: string
      responses:
        '201':
          description: Entertainment created successfully
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad request
  /api/v1/entertainments/{entertainment_id}:
    get:
      summary: Get entertainment details
      description: Retrieve detailed information about a specific entertainment
      operationId: get_entertainment
      tags:
        - entertainments
      parameters:
        - name: entertainment_id
          in: path
          required: true
          description: ID of the entertainment
          schema:
            type: integer
            minimum: 1
      responses:
        '201':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  entertainment_id:
                    type: integer
                  name:
                    type: string
                  city_id:
                    type: integer
                  event_time:
                    type: string
                    format: date-time
                  address:
                    type: string
                  cost:
                    type: number
                  description:
                    type: string
        '404':
          description: Entertainment not found
        '500':
          description: Internal server error

    put:
      summary: Update entertainment
      description: Update an existing entertainment and return HTML response
      operationId: update_entertainment
      tags:
        - entertainments
      parameters:
        - name: entertainment_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                city_id:
                  type: integer
                event_time:
                  type: string
                  format: date-time
                address:
                  type: string
                cost:
                  type: number
                description:
                  type: string
      responses:
        '201':
          description: Entertainment updated successfully
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Entertainment not found
        '500':
          description: Internal server error

    patch:
      summary: Partial update entertainment time
      description: Update only the event time of an entertainment
      operationId: update_entertainment_dates
      tags:
        - entertainments
      parameters:
        - name: entertainment_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                event_time:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Entertainment time updated successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Entertainment not found
        '500':
          description: Internal server error

    delete:
      summary: Delete entertainment
      description: Delete an entertainment and redirect to entertainments page
      operationId: delete_entertainment
      tags:
        - entertainments
      parameters:
        - name: entertainment_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '404':
          description: Entertainment not found
        '500':
          description: Internal server error
  /api/v1/route/{route_id}/entertainments/{entertainment_id}:
    delete:
      summary: Delete entertainment from route
      description: Delete an entertainment from a specific route and redirect to route edit page
      operationId: delete_entertainment_for_route
      tags:
        - route-entertainments
      parameters:
        - name: entertainment_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
        - name: route_id
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '404':
          description: Entertainment or route not found
        '500':
          description: Internal server error

    post:
      summary: Add entertainment to route
      description: Create a new entertainment and add it to a specific route
      operationId: add_ent_to_route
      tags:
        - route-entertainments
      parameters:
        - name: route_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                city_id:
                  type: integer
                event_time:
                  type: string
                  format: date-time
                address:
                  type: string
                cost:
                  type: number
                description:
                  type: string
      responses:
        '400':
          description: Bad request
        '404':
          description: Route not found
        '500':
          description: Internal server error

# ==================== СПРАВОЧНИК МАРШРУТОВ ====================
...
  /api/v1/travels:
    get:
      summary: Get all travels
      description: Retrieve all travels with users, entertainments and accommodations
      operationId: get_all_travels
      tags:
        - travels
      responses:
        '201':
          description: HTML page with travels data
          content:
            text/html:
              schema:
                type: string
        '500':
          description: Internal server error

    post:
      summary: Create new travel
      description: Create a new travel and return HTML response
      operationId: create_travel
      tags:
        - travels
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the travel
                description:
                  type: string
                  description: Description of the travel
                user_ids:
                  type: array
                  items:
                    type: integer
                  description: List of user IDs
                entertainment_ids:
                  type: array
                  items:
                    type: integer
                  description: List of entertainment IDs
                accommodation_ids:
                  type: array
                  items:
                    type: integer
                  description: List of accommodation IDs
      responses:
        '201':
          description: Travel created successfully
          content:
            text/html:
              schema:
                type: string
        '400':
          description: Bad request
  /api/v1/travels/{travel_id}:
    put:
      summary: Update travel
      description: Update an existing travel and return HTML response
      operationId: update_travel
      tags:
        - travels
      parameters:
        - name: travel_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                status:
                  type: string
                  enum: [planned, active, completed, cancelled]
                user_ids:
                  type: array
                  items:
                    type: integer
                entertainment_ids:
                  type: array
                  items:
                    type: integer
                accommodation_ids:
                  type: array
                  items:
                    type: integer
      responses:
        '201':
          description: Travel updated successfully
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Travel not found
        '500':
          description: Internal server error

    delete:
      summary: Delete travel
      description: Delete a travel and redirect to travels page
      operationId: delete_travel
      tags:
        - travels
      parameters:
        - name: travel_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '404':
          description: Travel not found
        '500':
          description: Internal server error

    patch:
      summary: Complete travel
      description: Mark a travel as completed and redirect to user profile
      operationId: complete_travel
      tags:
        - travels
      parameters:
        - name: travel_id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '404':
          description: Travel or user not found
          content:
            text/html:
              schema:
                type: string
                example: "<h1>Пользователь не найден</h1>"
        '500':
          description: Internal server error

# ==================== МАРШРУТ ====================
...